Implement API for obtaining glasses friendly name
This change allows the friendly name defined in the Tilt Five control panel to be accessed. 

This change also removes the default_display_name option which was producing  confusion because it was badly named and was misinterpreted for the the friendly name above. This was the name that was sent to the Tilt Five control panel to display with the glasses while the application was running. But really this should just be the application name and so the plugin now just sends the application name in it's place.

The example apps were also amended to demo the feature.

Co-authored-by:  Jonathan Stevens <jstevens@tiltfive.com>